Db final. Fixed arthropod info not properly recorded.

+++ www/js/input-page.js
@@ -516,6 +516,7 @@ var saveArthropod = function( ) {
 		//);
 
 	}
+	submitArthropodsToDB(time,selectedOrder,length,count,notes);
 };
 
 
@@ -648,7 +649,6 @@ var submit = function( ) {
 
 	if(online == false){
         //last field for error handler 0 is default
-		submitArthropodsToDB(dateTime);
 		//alert("I am here");
 		db.transaction(function(tx){
                         tx.executeSql("INSERT INTO SURVEY VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)", 
@@ -673,19 +673,21 @@ var submit = function( ) {
                         alert("Transaction Error: "+error.message);
                     },function(){
 						alert("This page was successfully stored");
-						db.transaction(function(tx){
-            					tx.executeSql("DELETE from SURVEY where timeStart=?", [timeStart]);
-        				},  function(error){
-            				alert("Transaction error: "+error.message);
-        				}, function(){
-            				//alert("Successfully delete this survey");
-        				});
 						window.location = "homepage.html";
 
 		});
 			
 	}else{
 		submitSurveyToServer();	
+	}
+	if(edit){
+		db.transaction(function(tx){
+            					tx.executeSql("DELETE from SURVEY where timeStart=?", [timeStart]);
+        				},  function(error){
+            				alert("Transaction error: "+error.message);
+        				}, function(){
+            				//alert("Successfully delete this survey");
+        				});
 	}
 
 
@@ -799,12 +801,10 @@ var submitSurveyToServer = function(){
 };
 
 
-var submitArthropodsToDB = function(time){
+var submitArthropodsToDB = function(time,selectedOrder,length,count,notes){
 	var arthropodInputs = $(".arthropod-input");
 	numberOfArthropodsToSubmit = arthropodInputs.length;
 	numberOfArthropodsSubmitted = 0;
-	if(numberOfArthropodsToSubmit > 0) {
-		arthropodInputs.each(function () {
 
 			//Get values of caterpillar checklist
 			var hairyOrSpiny, leafRoll, silkTent;
@@ -831,10 +831,10 @@ var submitArthropodsToDB = function(time){
 			//navigator.notification.alert("Arthropod image uri: " + arthropodImageURI);
            db.transaction(function(tx){
                         tx.executeSql("INSERT INTO ARTHROPODS VALUES (?,?,?,?,?,?,?,?,?)", 
-                        	[$("h4", this).text(),
-                        	parseInt($(".arthropod-length", this).text()),
-                        	$(".arthropod-notes", this).text(),
-                        	parseInt($(".arthropod-count", this).text()),
+                        	[selectedOrder,
+                        	length,
+                        	notes,
+                        	count,
                         	hairyOrSpiny,
                         	leafRoll,
                         	silkTent,
@@ -843,7 +843,9 @@ var submitArthropodsToDB = function(time){
                     }  , function(error){
                         alert("Transaction Error: "+error.message);
                     },function(){
-						alert("Arthropods information was successfully stored");
+						//alert(selectedOrder);
+						//alert(length);
+						//alert(count);
 						db.transaction(function(tx){
             					tx.executeSql("DELETE from SURVEY where timeStart=?", [timeStart]);
         				},  function(error){
@@ -855,13 +857,9 @@ var submitArthropodsToDB = function(time){
 
 
 
-		});
-	}
-	else{
 		navigator.notification.alert("Successfully submitted survey data!");
 		//clearFields();
-	}
-
+	
 };
